videoDownloadProject/requirements/dev.txt
-r base.txt
django-browser-reload==1.21.0
black==26.1.0
isort==7.0.0
ruff==0.15.0
djlint==1.36.4
